Fix useless cast warning

+++ include/boost/decimal/decimal32_t.hpp
@@ -1075,7 +1075,7 @@ BOOST_DECIMAL_CUDA_CONSTEXPR auto isnormal BOOST_DECIMAL_PREVENT_MACRO_SUBSTITUT
     // A value is subnormal when its adjusted exponent falls below emin. That depends on how
     // many digits the significand carries, not on the quantum exponent alone: 1E-95 and
     // 1.000000E-95 are both the smallest normal decimal32_t but sit 6 quanta apart.
-    const auto adjusted_exp {static_cast<int>(rhs.biased_exponent()) + detail::num_digits(sig) - 1};
+    const auto adjusted_exp {rhs.biased_exponent() + detail::num_digits(sig) - 1};
 
     return adjusted_exp >= detail::emin_v<decimal32_t>;
     #else
