+++ include/boost/decimal/decimal64_t.hpp
@@ -1322,7 +1322,7 @@ BOOST_DECIMAL_CUDA_CONSTEXPR auto isnormal BOOST_DECIMAL_PREVENT_MACRO_SUBSTITUT
     // A value is subnormal when its adjusted exponent falls below emin. That depends on how
     // many digits the significand carries, not on the quantum exponent alone: 1E-383 and
     // 1.000000000000000E-383 are both the smallest normal decimal64_t but sit 15 quanta apart.
-    const auto adjusted_exp {static_cast<int>(rhs.biased_exponent()) + detail::num_digits(sig) - 1};
+    const auto adjusted_exp {rhs.biased_exponent() + detail::num_digits(sig) - 1};
 
     return adjusted_exp >= detail::emin_v<decimal64_t>;
     #else
